llm-api/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
transformers
torch
accelerate
sentencepiece